Changed QueryString to NonDelimitedText and updated tests
